tests/requirements.txt
# Test requirements for News Intelligence System v4.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
requests>=2.28.0
psycopg2-binary>=2.9.0
pytest-html>=3.1.0
pytest-json-report>=1.5.0
